Moved Sidebar component to a folder
